+++ app/src/app/campaign/based-hanabi/artworks/[id]/page.tsx
@@ -1,0 +1,3 @@
+export default function CampaignBasedHanabiArtworkPage() {
+  return <>CampaignBasedHanabiPage</>;
+}
