+++ app/src/app/campaign/based-hanabi/artworks/create/page.tsx
@@ -2,7 +2,7 @@
 
 import PixelEditor from "@/components/PixelEditor";
 
-export default function TestEditorPage() {
+export default function CampaignBasedHanabiArtworkCreatePage() {
   return (
     <div className="">
       <PixelEditor />
