+++ app/src/app/campaign/based-hanabi/page.tsx
@@ -1,0 +1,3 @@
+export default function CampaignBasedHanabiPage() {
+  return <>CampaignBasedHanabiPage</>;
+}
